requirements.txt
# Django Core
Django==6.0.3
asgiref==3.11.1
sqlparse==0.5.5

# GraphQL Ecosystem
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.8
graphql-relay==3.2.0

# CORS Support
django-cors-headers==4.9.0

# Django Filters
django-filter==25.2

# Image Processing
Pillow==12.1.1

# Environment Variables
python-dotenv==1.2.2

# GraphQL Dependencies
promise==2.3
python-dateutil==2.9.0.post0
six==1.17.0
text-unidecode==1.3
typing_extensions==4.15.0

# SQLite Support (built into Django, no additional package needed)

# WebSockets Support (إذا كنت تحتاجها)
channels==3.0.4
channels-redis==3.4.1
daphne==3.0.2
redis==3.5.3
websockets==12.0
uvicorn[standard]==0.24.0

# Background Tasks
django-q==1.3.9

# JWT Support
PyJWT==2.12.1
